// tools/importers/blue-crest/types.ts
export interface ImportedUnit {

  unitNumber: string;

  building: string;

  level: number;

  unitType: string;

  bedrooms: number;

  area: number;

  view: string;

  currency: string;

  price: number;

  downPayment: number;

  quarterly: number;

  maintenance: number;

  facilities: number;

}